suzuran: Link voice-dmic-ef directly to handset-dmic-endfire
Linking the incall dual mic to the mono speaker mic wasn't a good idea since it only echoed the your own voice rather than giving output on the other side. Thankfully i've found what seems to be a solution by directly linking voice-dmic-ef to handset-dmic-endfire. Oddly enough the standart configuration which links voice-dmic-ef to dmic-endfire which links to handset-dmic-endifre should give identical results. A theory would be that the stock proprietary HAL links the configuration to another mixer path rather than dmic-ef and eventually to handset-dmic-endfire or directly to handset-dmic-endfire as we did. However to get an absolute solution for this issue a source code of the proprietary HAL is required.

-CTXz